fix 64 bits time_t issues

+++ src/driver/implementation/mh/mhdriver_cached_message.c
@@ -187,8 +187,8 @@ static int mh_initialize(mailmessage * msg_info)
   
   mh_msg_info = data.data;
 
-  snprintf(static_uid, PATH_MAX, "%u-%lu-%lu", msg_info->msg_index,
-	   mh_msg_info->msg_mtime, (unsigned long) mh_msg_info->msg_size);
+  snprintf(static_uid, PATH_MAX, "%u-%lld-%zu", msg_info->msg_index,
+           (long long)mh_msg_info->msg_mtime, mh_msg_info->msg_size);
   uid = strdup(static_uid);
   if (uid == NULL)
     return MAIL_ERROR_MEMORY;
